Security fixes

+++ deploy.sh
@@ -5,6 +5,8 @@ LN_STACK=$2
 STACK_NAME=$3
 BASEDIR=$(pwd)
 
+NODEVERSION=$(node --version)
+
 if [ "$BUCKET" == "" ] ;
 then
   echo "Bucket name is missing"
@@ -21,6 +23,12 @@ then
   STACK_NAME=$BUCKET;
 fi
 
+if [ ! "$NODEVERSION" -eq "v6.10.1" ]
+then
+    echo "Node version must be 6.10.1"
+    exit 1
+fi
+
 for d in ./functions/* ;
 do
     cd $BASEDIR/$d && rm -rf node_modules && npm install --production ;
@@ -28,5 +36,12 @@ done
 
 cd $BASEDIR;
 
-aws cloudformation package --template-file cloudformation.json --s3-bucket $BUCKET --output-template-file output.yml &&\
-  aws cloudformation deploy --template-file output.yml --stack-name $STACK_NAME --capabilities CAPABILITY_IAM --parameter-overrides LNStack=$LN_STACK
+aws cloudformation package \
+    --template-file cloudformation.json \
+    --s3-bucket $BUCKET \
+    --output-template-file output.yml \
+    && aws cloudformation deploy \
+        --template-file output.yml \
+        --stack-name $STACK_NAME \
+        --capabilities CAPABILITY_IAM \
+        --parameter-overrides LNStack=$LN_STACK
